Create separate calendar component
